# openapi.yaml
openapi: 3.0.3
info:
  title: Strategic Planning Platform API
  description: |
    AI-Powered Strategic Planning Platform with GraphRAG validation.
    
    ## Features
    
    * **4-Phase PRD Creation Workflow**: From concept to comprehensive document
    * **GraphRAG Validation**: <2% hallucination rate through three-tier validation
    * **PydanticAI Agents**: Intelligent PRD processing and task generation
    * **GitHub Integration**: Automated project setup and task management
    * **Enterprise Security**: JWT auth, RBAC, and comprehensive audit logging
    
    ## Performance Targets
    
    * API Response Time: <200ms P95
    * Document Generation: <60 seconds
    * Concurrent Users: 100+ with stable performance
    * Uptime: 99.9% SLA
  version: 1.0.0
  contact:
    name: Strategic Planning Platform Team
    email: support@strategicplanning.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.strategicplanning.ai
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /register:
    post:
      tags:
        - Authentication
      summary: Register new user account (alias)
      description: Create a new user account with email and password
      operationId: registerUserAlias
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Registration failed - invalid input or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user account
      description: Create a new user account with email and password
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Registration failed - invalid input or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /login:
    post:
      tags:
        - Authentication
      summary: User login (alias)
      description: Authenticate user and return JWT tokens
      operationId: loginUserAlias
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      description: Retrieve current authenticated user information
      operationId: getCurrentUser
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token (alias)
      operationId: refreshTokenAlias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - status_code
        - timestamp
      properties:
        error:
          type: string
          description: Error message
        status_code:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: Request path
        request_id:
          type: string
          description: Request correlation ID
...
    UserRegister:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password
        full_name:
          type: string
          minLength: 2
          maxLength: 100
          description: User full name
        company:
          type: string
          maxLength: 100
          description: Company name
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        token_type:
          type: string
          default: bearer
          description: Token type
        expires_in:
          type: integer
          description: Token expiration in seconds
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required:
        - id
        - email
        - full_name
        - role
        - is_active
        - created_at
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
          description: User email
        full_name:
          type: string
          description: User full name
        company:
          type: string
          description: Company name
        role:
          type: string
          default: user
          description: User role
        is_active:
          type: boolean
          default: true
          description: User active status
        created_at:
          type: string
          format: date-time
          description: Account creation date
...